+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 
 ##### Release date: 31/5/2022
 - improvement: format api request function
+- fix-bug: delete mutil error